requirements.txt
agno>=0.1.0
supabase>=2.0.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
pydantic>=2.0.0
